Move `resetOtp` to the right file
