+++ src/core/context/context-api.js
@@ -19,7 +19,8 @@ import { requestEdgeLogin } from '../login/edge.js'
 import { fetchLoginMessages } from '../login/login-messages.js'
 import { getStash } from '../login/login-selectors.js'
 import { removeStash } from '../login/login-stash.js'
-import { makeLoginTree, resetOtp, syncLogin } from '../login/login.js'
+import { makeLoginTree, syncLogin } from '../login/login.js'
+import { resetOtp } from '../login/otp.js'
 import { loginPassword } from '../login/password.js'
 import { getPin2Key, loginPin2 } from '../login/pin2.js'
 import {
